+++ ostap/io/sqliteshelve.py
@@ -347,7 +347,7 @@ class SQLiteShelf(SqliteDict):
     #  db  = ...
     #  ndb = db.clone ( 'new_file.db' )
     #  @endcode
-    def clone ( self , new_name ) : 
+    def clone ( self , new_name , keys = () ) : 
         """ Clone the database into new one
         >>> old_db = ...
         >>> new_db = new_db.clone ( 'new_file.db' )
@@ -360,8 +360,12 @@ class SQLiteShelf(SqliteDict):
                                compress_level = self.compression  , 
                                journal_mode   = self.journal_mode )
         
-        ## copy the content 
-        for key in self.keys() : new_db [ key ] = self [ key ]
+        ## copy the content
+        if keys :
+            for key in self.keys() :
+                if key in keys     : new_db [ key ] = self [ key ]
+        else : 
+            for key in self.keys() : new_db [ key ] = self [ key ]
         
         new_db.sync ()  
         return new_db